Move audio channel things to its own class
